sia10: add 'base' coordinate frame to xacro. For #45.
This corresponds to the Motoman 'Robot Frame' (not the 'Base Frame'). This
frame has its origin at the intersection of the S axis and a horizontal
plane going through the L axis.

Note: the frame was added manually to the urdf, so the '.urdf' is still out
of sync with the xacro macro (issue #107).