MotoROS: refactor IO server into stand-alone task.
This migrates all code dealing with IO (reading, writing) to a separate task,
instead of multiplexing it into and out of the MotionServer datastreams.

A new (TCP) server is added on TCP:50242, supporting a maximum of one (1)
connection.

The IO task runs a periodic while-loop at 50 Hz, at the same priority as the
State and Motion server tasks.

Finally, a new IO feedback signal is defined (nr 11126), taking the place of
IO_FEEDBACK_RESERVED_0.

Note: this will need support in motoman_driver. That is not included in this
commit.