driver: remove unused overload of withinGoalConstraints(..).
Overload is not used by any part of the JTA.

Additionally, it was checking only the status of the fourth motion group, which is not a correct way to determine motion completion for the entire robot (at least not until 'in_motion' was fixed in #227).

If needed, we can bring it back by reverting this commit.