Correction on urdf and model orientation and direction of motion.
The URDF was changed to rotate the whole model so that the X-axis point forward and kinematic model adjusted to match the standard Yaskawa model and dimensions.  The mesh were then adjusted to realign properly with the new kinematic model.  The direction of motion for each axis was also corrected to match the real robot motion.  Finally, the visual meshes were reduce by 50% to reduce the display processing.