Improve PFL handling in MotoROS.
Summary:

 - Fixed motion_possible status when PFL is active
 - Added PFL Avoidance signal
 - Added PFL flag to prevent continuing previous motion after PFL activation

Detailed description:

MotoROS PFL Handling
--------------------

Following the bug report #338 "Robot status motion possible flag incorrect when robot enable called during PFL stop" on the github motoman repo, the MotoROS PFL handling was review.
It was found that if the robot is in PFL stop, and the `/robot_enable` service is called, the `motion_possible` flag in the `/robot_state` messages is set to `true`, but the robot cannot be moved.
Furthermore, the PFL Avoidance mode was not originally covered.
Under some PFL modes, some motion could occur after PFL activation completed but that motion was no longer synchronized with the desired ROS motion.

PFL (Power Force Limiting) Overview
-----------------------------------

The PFL (Power Force Limiting) function of the HC-series robots has 3 modes of operations:

- Transient Contact: Temporary contact caused by the robot hits something, or something hits the robot. Robot stops and the amber light button on the upper arm turns on. When the amber light button is press, the light turns off and the motion is resumes.

- Quasi-Static Contact: Similar to the transient contact but the contact is maintained, like hitting a table or a wall. After stopping, if an external force is maintained, the robot will move away from the exerted force to release the pressure and then stop again. The amber light button on the upper arm turns back on. When the amber light button is press, the light turns off and the motion is resumed.

- Avoidance: A low external force is applied to move the robot from its current trajectory. Once the force is removed, the robot will move back to the position where the external force was originally applied. Then the robot resumes its original path.

Note that some modes may be disabled on some systems.

ROS Handling of PFL
-------------------

The PFL function, being a safety feature, supersedes any motion from ROS.
The PFL normal behavior should be allowed to take place and ROS should not interfere with it by sending stop motion while the PFL function is active.
When ROS is moving the robot, since the motion path is coming from an external source and the position of the robot may have changed following the activation of PFL, MotoROS will cancel the current trajectory following the PFL activation.
Following the PFL activation any joint trajectory message sent to the controller will return the `ROS_RESULT_NOT_READY_PFL_ACTIVE` code `5011`.
The status bit `in_motion` will be `false` when the robot is not moving.
It may temporarily turn on in PFL's *Quasi-Static Contact* mode as the robot moves away to release pressure on the impact point.
The status bit `motion_possible` (ready) will be `false` and remain `false` until the `/robot_enable` service is called (Start trajectory mode message `ROS_MSG_MOTO_MOTION_CTRL` message `2001` with the `ROS_CMD_START_TRAJ_MODE` command `200121`).
When `motion_possible` is `true` again, the current robot position can be acquired by ROS and a new trajectory to resume motion can be planned and sent to the controller.

Code Change
-----------

Code change essentially consists of adding new I/O signals to capture the avoidance mode:

- `IO_ROBOTSTATUS_PFL_AVOIDING` (15120)
- `IO_ROBOTSTATUS_PFL_AVOID_JOINT` (15124)
- `IO_ROBOTSTATUS_PFL_AVOID_TRANS` (15125)

The `IO_ROBOTSTATUS_PFL_AVOIDING` signal indicates that a force above the avoidance threshold was detected.
The signals `IO_ROBOTSTATUS_PFL_AVOID_JOINT` and `IO_ROBOTSTATUS_PFL_AVOID_TRANS` only indicate that the avoidance mode is enabled.
So, the `IO_ROBOTSTATUS_PFL_AVOIDING` signal should only be considered if one of the other two signals is on.

And a new flag `bool bPFLduringRosMove` was added to the `Controller` class to track that PFL was activated.
Finally, logic to manage the detection and reset of the PFL flag was added.