driver: lint only specific files.
instead of using a CPPLINT.cfg file to exlude files from the set found by the roslint_cpp(..) macro, we do our own globbing and then remove the files we don't want cpplint to see.

This avoids the need for a CPPLINT.cfg file in the MotoPlus source tree (as cpplint requires that file in the exact directory the options should apply to).

We deem globbing acceptable here (and are aware of the disadvantages), as this is already done by roslint_cpp(..), so we're not introducing those disadvantages with this commit.