driver: log (more) appropriate msgs when topic_list is not found
In most cases this is not necessarily a problem, as single-group systems fi are not expected to not have that parameter.

So log an INFO message and continue trying to read the single-group config from the 'controller_joint_names' parameter (if that can't be found, we have a good reason to complain / refuse to continue).

Keep the WARN for one particular case: users starting the multi-group version of the JTA, but then *not* correctly setting the 'topic_list' parameter should be warned.

Thanks to @cjue who proposed the initial fix (in #483), which I've slightly expanded upon here.